git-svn-id: https://plugins.svn.wordpress.org/sitepress-multilingual-cms/trunk@237381 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/sitepress-schema.php
@@ -327,18 +327,20 @@ function icl_sitepress_activate(){
     }
     
 	global $wp_taxonomies;
-	foreach ($wp_taxonomies as $t => $v) {
-		$orphans = $wpdb->get_col("
+	if (is_array($wp_taxonomies)) {
+		foreach ($wp_taxonomies as $t => $v) {
+			$orphans = $wpdb->get_col("
 		SELECT t.translation_id 
 		FROM {$wpdb->prefix}icl_translations t 
         LEFT JOIN {$wpdb->term_taxonomy} p 
 		ON t.element_id = p.term_taxonomy_id 
 		WHERE t.element_type = 'tax_{$t}' 
 		AND p.taxonomy <> '{$t}'
-		");
-    	if(!empty($orphans)){
-        	$wpdb->query("DELETE FROM {$wpdb->prefix}icl_translations WHERE translation_id IN (".join(',',$orphans).")");
-    	}
+			");
+    		if (!empty($orphans)) {
+        		$wpdb->query("DELETE FROM {$wpdb->prefix}icl_translations WHERE translation_id IN (".join(',',$orphans).")");
+    		}
+		}
 	}
 	
     if(defined('ICL_DEBUG_MODE') && ICL_DEBUG_MODE){
